Order mode: X button fix + remove on double click
The delete button (X) was disabled by being located under an event box
in some situations (fixed by showing the button in a high priority
idle funktion after the event box establishes "above" mode, but before
redraw events - this is sort of a hack..).

Any other place of the bar can be double clicked now to remove the
effect from the list (in order mode).